+++ .npmignore
@@ -1,4 +1,4 @@
 /.idea
-/src/rocketLcpBeacon.js
+/src/RocketFinalBeacon.js
 /test
 /.github
