--- src/rocketLcpBeacon.js
@@ -1,22 +1,0 @@
-import LcpBeacon from './LcpBeacon.js';
-
-( rocket_lcp_data => {
-    if ( !rocket_lcp_data ) {
-        return;
-    }
-
-    const instance = new LcpBeacon( rocket_lcp_data );
-
-    if (document.readyState !== 'loading') {
-        setTimeout(() => {
-            instance.init();
-        }, rocket_lcp_data.delay);
-        return;
-    }
-
-    document.addEventListener("DOMContentLoaded", () => {
-        setTimeout(() => {
-            instance.init();
-        }, rocket_lcp_data.delay);
-    });
-} )( window.rocket_lcp_data );
